Refactor objs and attrs

+++ src/skrough/typing_utils.py
@@ -4,8 +4,8 @@ import skrough.typing as rght
 
 
 def unify_objs(objs: rght.ObjsLike) -> rght.Objs:
-    return np.asarray(objs)
+    return np.asarray(objs, dtype=np.int64)
 
 
 def unify_attrs(attrs: rght.AttrsLike) -> rght.Attrs:
-    return np.asarray(attrs)
+    return np.asarray(attrs, dtype=np.int64)
